move interface to modules/manager
