+++ lib/workers/repository/extract/index.ts
@@ -8,9 +8,11 @@ import type {
 } from '../../../config/types';
 import { logger } from '../../../logger';
 import { getManagerList } from '../../../modules/manager';
-import type { PackageFile } from '../../../modules/manager/types';
+import type {
+  PackageFile,
+  WorkerExtractConfig,
+} from '../../../modules/manager/types';
 import { getFileList } from '../../../util/git';
-import type { WorkerExtractConfig } from '../../types';
 import { getMatchingFiles } from './file-match';
 import { getManagerPackageFiles } from './manager-files';
