+++ lib/workers/repository/extract/manager-files.ts
@@ -5,9 +5,11 @@ import {
   extractPackageFile,
   get,
 } from '../../../modules/manager';
-import type { PackageFile } from '../../../modules/manager/types';
+import type {
+  PackageFile,
+  WorkerExtractConfig,
+} from '../../../modules/manager/types';
 import { readLocalFile } from '../../../util/fs';
-import type { WorkerExtractConfig } from '../../types';
 
 export async function getManagerPackageFiles(
   config: WorkerExtractConfig
