+++ lib/workers/types.ts
@@ -2,7 +2,6 @@ import type { Merge } from 'type-fest';
 import type {
   GroupConfig,
   LegacyAdminConfig,
-  RegExManager,
   RenovateConfig,
   RenovateSharedConfig,
   ValidationMessage,
@@ -10,8 +9,6 @@ import type {
 import type { Release } from '../modules/datasource/types';
 import type {
   ArtifactError,
-  CustomExtractConfig,
-  ExtractConfig,
   LookupUpdate,
   PackageDependency,
   PackageFile,
@@ -125,18 +122,4 @@ export interface BranchConfig
   prNo?: number;
   stopUpdating?: boolean;
   isConflicted?: boolean;
-}
-
-export interface WorkerExtractConfig
-  extends ExtractConfig,
-    Partial<CustomExtractConfig> {
-  manager: string;
-  fileList?: string[];
-  fileMatch?: string[];
-  updateInternalDeps?: boolean;
-  includePaths?: string[];
-  ignorePaths?: string[];
-  regexManagers?: RegExManager[];
-  enabledManagers?: string[];
-  enabled?: boolean;
 }
